+++ projectroles/forms.py
@@ -1033,6 +1033,13 @@ class ProjectUserCreateForm(SODARModelForm):
             label='Confirm password', widget=forms.PasswordInput()
         )
 
+    def save(self, commit=True):
+        user = super().save(commit=False)
+        user.set_password(self.cleaned_data['password'])
+        if commit:
+            user.save()
+        return user
+
     def clean(self):
         cleaned_data = super().clean()
         if cleaned_data['password'] != cleaned_data['password_confirm']: